Migration to use lst_utilities.py
